demo/backend/requirements.txt
Flask
python-dotenv
pyjwt
requests
flask-cors
redisvl
redis
google-cloud-aiplatform
